+++ build.gradle
@@ -37,9 +37,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-mail:2.7.1'
 //카카오
     implementation'com.google.code.gson:gson:2.8.7'
-
-    implementation group: 'org.apache.httpcomponents', name: 'httpclient', version: '4.5.3'
-    implementation group: 'com.googlecode.json-simple', name: 'json-simple', version: '1.1'
+//네이버
 
 }
