+++ src/main/java/com/heydoctor/app/HeydocterApplication.java
@@ -1,5 +1,6 @@
 package com.heydoctor.app;
 
+import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
