+++ src/main/java/com/heydoctor/app/domain/vo/UserVO.java
@@ -22,4 +22,5 @@ public class UserVO {
     private String userProfileType;
     private String userProfileUuid;
 
+
 }
