+++ src/main/resources/application.yml
@@ -1,11 +1,9 @@
-#server
 server:
   port: 10000
   servlet:
     session:
       tracking-modes: cookie
 
-#MyBatis
 spring:
   datasource:
     hikari:
@@ -14,7 +12,6 @@ spring:
       username: hr
       password: hr
 
-#Multipart
   servlet:
     multipart:
       enabled: true
