+++ src/main/resources/mapper/UserMapper.xml
@@ -18,11 +18,15 @@
         WHERE USER_ID = #{userId}
     </select>
 
-
     <update id="kakaoUpdate" parameterType="userVO">
-    UPDATE TBL_USER
-    SET USER_EMAIL = #{userEmail}, USER_NAME = #{userName}, USER_PASSWORD = #{userPassword}
-    WHERE USER_EMAIL = #{userEmail}
-</update>
+        UPDATE TBL_USER
+        SET USER_EMAIL = #{userEmail}, USER_NAME = #{userName}, USER_PASSWORD = #{userPassword}
+        WHERE USER_EMAIL = #{userEmail}
+    </update>
 
+    <update id="naverUpdate" parameterType="userVO">
+        UPDATE TBL_USER
+        SET USER_EMAIL = #{userEmail}, USER_NAME = #{userName}, USER_PASSWORD = #{userPassword}
+        WHERE USER_EMAIL = #{userEmail}
+    </update>
 </mapper>
