+++ src/main/resources/static/css/login-page/login-start.css
@@ -130,10 +130,6 @@ div {
     display: block;
 }
 
-.ml-6 {
-    margin-left: 6px!important;
-}
-
 .c-body1 {
     font-size: 14px;
     line-height: 2px;
