+++ src/main/resources/templates/login/callback.html
@@ -1,26 +1,47 @@
 <!doctype html>
 <html lang="ko">
 <head>
-    <script type="text/javascript" src="https://static.nid.naver.com/js/naverLogin_implicit-1.0.3.js"
-            charset="utf-8"></script>
+    <script type="text/javascript" src="https://static.nid.naver.com/js/naverLogin_implicit-1.0.3.js" charset="utf-8"></script>
     <script type="text/javascript" src="http://code.jquery.com/jquery-1.11.3.min.js"></script>
 </head>
 <body>
-callback
-<script type="text/javascript">
-    var naver_id_login = new naver_id_login("0QEJy5lMTxel75jSVhqm", "http://localhost:10000/login/naver");
-    // 접근 토큰 값 출력
-    alert(naver_id_login.oauthParams.access_token);
-    // 네이버 사용자 프로필 조회
-    naver_id_login.get_naver_userprofile("naverSignInCallback()");
-
-    // 네이버 사용자 프로필 조회 이후 프로필 정보를 처리할 callback function
-    function naverSignInCallback() {
-        console.log(naver_id_login.getProfileData('email'));
-        console.log(naver_id_login.getProfileData('name'));
-        console.log(naver_id_login.getProfileData('id'));
-    }
-
-</script>
 </body>
 </html>
+<script type="text/javascript">
+    var naver_id_login = new naver_id_login("0QEJy5lMTxel75jSVhqm", "http://localhost:10000/login/naver");
+    // 네이버 사용자 프로필 조회 이후 프로필 정보를 처리할 callback function
+    console.log(naver_id_login.oauthParams.access_token);
+    function naverSignInCallback() {
+        var email = naver_id_login.getProfileData('email');
+        var name = naver_id_login.getProfileData('name');
+        var id = naver_id_login.getProfileData('id');
+
+        console.log(naver_id_login.getProfileData('email'));
+        console.log(naver_id_login.getProfileData('name'));
+        console.log(naver_id_login.getProfileData('identification'));
+
+        // 이메일 값이 비어 있지 않은 경우에만 서버로 데이터 전송
+
+        $.ajax({
+            url: '/login/navera',
+            type: 'GET',
+            data: {
+                email: email,
+                name: name,
+                id: id
+            },
+            success: function(response) {
+                console.log('프로필 데이터가 성공적으로 저장되었습니다.');
+                location.replace(response);
+
+            },
+            error: function(error) {
+                console.error('프로필 데이터 저장 중 오류가 발생했습니다:', error);
+            }
+        });
+
+    }
+
+    // 네이버 사용자 프로필 조회
+    naver_id_login.get_naver_userprofile("naverSignInCallback()");
+</script>
